requirements.txt
telethon
python-dotenv
azure-ai-inference
MetaTrader5
